+++ src/main/java/ir/mctracker/core/commands/TrackerCommand.java
@@ -1,8 +1,8 @@
-package ir.mctracker.commands;
+package ir.mctracker.core.commands;
 
 import ir.mctracker.MCTrackerVote;
-import ir.mctracker.config.Messages;
-import ir.mctracker.utilities.Util;
+import ir.mctracker.core.config.Messages;
+import ir.mctracker.core.utilities.Util;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
