+++ src/main/java/ir/mctracker/core/database/DataSource.java
@@ -1,4 +1,4 @@
-package ir.mctracker.database;
+package ir.mctracker.core.database;
 
 import com.j256.ormlite.dao.DaoManager;
 import com.j256.ormlite.jdbc.JdbcConnectionSource;
@@ -6,7 +6,7 @@ import com.j256.ormlite.support.ConnectionSource;
 import com.j256.ormlite.table.TableUtils;
 import ir.jeykey.megacore.config.premade.Storage;
 import ir.mctracker.MCTrackerVote;
-import ir.mctracker.database.models.Vote;
+import ir.mctracker.core.database.models.Vote;
 import lombok.Getter;
 
 import java.io.File;
