+++ src/main/java/ir/mctracker/core/database/models/Vote.java
@@ -1,4 +1,4 @@
-package ir.mctracker.database.models;
+package ir.mctracker.core.database.models;
 
 import com.j256.ormlite.field.DatabaseField;
 import com.j256.ormlite.table.DatabaseTable;
