+++ src/main/java/ir/mctracker/core/utilities/Metrics.java
@@ -1,4 +1,4 @@
-package ir.mctracker.utilities;
+package ir.mctracker.core.utilities;
 
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.YamlConfiguration;
