+++ src/main/java/ir/mctracker/core/utilities/Util.java
@@ -1,6 +1,6 @@
-package ir.mctracker.utilities;
+package ir.mctracker.core.utilities;
 
-import ir.mctracker.config.Messages;
+import ir.mctracker.core.config.Messages;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
